refactored crypto to a new file.
